move toward more consistent naming
